Adds final solution
This was a pretty simple game, so I just made it in one go and finished
instead of making several smaller commits.

This includes the bare html file and the javascript file where all the
code for the game is. In future versions, I need to add a better user
interface.