Implements UI
This allows the user to select their action instead of type it which
greatly simplifies the code since I don't need to check on mistyped
actions or wrong capitalization. All the basic function should be done
with it. I just need to add some style and make sure that things
are lined up correctly.

This creates the styles css file, but I have not had the time to add
the styles into the file. That will be done at a later time.